backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
sentence-transformers
faiss-cpu
groq
python-dotenv
numpy